fix(inheritance-report): Debts and funeral cost
