+++ libs/application/templates/inheritance-report/src/types.ts
@@ -210,12 +210,11 @@ export interface BankAccounts {
   total: number
 }
 
-export interface AllDebts {
-  balance: string
+export interface Debt {
+  assetNumber: string
   nationalId: string
-  creditorName: string
-  loanIdentity: string
-  taxFreeInheritance: number
+  description: string
+  propertyValuation: string
 }
 
 export interface ApplicationDebts {
@@ -232,7 +231,7 @@ interface DomesticAndForeignDebtsData {
 }
 
 interface DomesticAndForeignDebts {
-  data: DomesticAndForeignDebtsData[]
+  data: Debt[]
   total: number
 }
